+++ tools/src/html-loader/loader.ts
@@ -38,8 +38,7 @@ showdown.extension('prism', () => {
   ];
 });
 
-export default function (markdown: string) {
-
+export default function (markdown) {
   this.cacheable();
 
   const converter = new showdown.Converter({